Feat: Adds .stl creation and retrieval pipeline
This commit adds a new `manager` for roost that takes a user's request, generates the required scad code, runs this code and returns the corresponding .stl.

It has been called a `manager` because it assigns unique identifiers to each instance of a model .stl based on the parameters requested, allowing for caching of .stl files and avoidance of duplicates. In the future it will also manage how large this cache will get and prune it based on individual file `popularity`.