Feat: New UI layout and proper camera placement
In light of my previous UI 'fix' I decided that the old layout was flawed, so this new layout is based almost purely on the MUI grid layout. It plays a little nicer with the viewport-filling style.

This also dynamically places the Threejs camera based on the model's bounding box and adds a camera icon that allows the user to reset its positioning.