Fix: Makes all parameter changes generate a new stl
Previously the ListRestriction and input fields from ParameteElements.js were broken because the changes would not take effect until the next render call. This happened because the formValues changed too slowly and were missed on the 'current' render call.

This change should fix this bug, but if I'm honest, I'm not entirely convinced yet.