Feat: Updates roost to use the database as the source of info
Instead of the 'index.json' file that was previously outputted by plume, now the database is used. This is a marked improvement in many ways.

However, the current implementation is slow because each database related function acquires a new database connection, even when it is not strictly necessary. This is a product of my laziness to figure out the correct borrowing. Will fix later.