Feat: Adds .stl instance restoration function
If the '--restore' flag is used with 'plume index', plume will attempt to restore the .stl instances wiped by the 'npm run build' command.

This is useful because it will mitigate the slower response times associated with generating a new model in parakeet (at least for the instances that are successfully restored).